migrations/models/4_20241016164608_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `task` MODIFY COLUMN `endTime` DATETIME(6);"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `task` MODIFY COLUMN `endTime` DATETIME(6) NOT NULL;"""
