requirement.txt
fastapi[all]==0.75.2
aerich
tortoise-orm[aiomysql]==0.19.0
python-multipart

